Fix a broken test in ImagesErrorsTest
