+++ api/api/src/test/scala/uk/ac/wellcome/platform/api/images/ImagesErrorsTest.scala
@@ -1,5 +1,7 @@
 package uk.ac.wellcome.platform.api.images
 
+import org.scalatest.Assertion
+
 class ImagesErrorsTest extends ApiImagesTestBase {
   describe("returns a 404 for missing resources") {
     it("looking up an image that doesn't exist") {
@@ -10,4 +12,16 @@ class ImagesErrorsTest extends ApiImagesTestBase {
       )
     }
   }
+
+  def assertIsNotFound(path: String, description: String): Assertion =
+    withImagesApi {
+      case (_, routes) =>
+        assertJsonResponse(routes, s"/$apiPrefix$path")(
+          Status.NotFound ->
+            notFound(
+              apiPrefix = apiPrefix,
+              description = description
+            )
+        )
+    }
 }
